requirements.txt
django
djangorestframework
mysql-connector-python
django-cors-headers
djangorestframework_simplejwt
drf-yasg
python-dotenv
bcrypt